generated/segment_validation/segment-validation__t2.html

<!DOCTYPE html>
<html lang="en">
  <head>
    <meta charset="utf-8" />
    <title>Segment Validation</title>
    <link rel="stylesheet" href="../lib/style.css" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>

  <body>
    <div id="content-wrapper">
      <div id="video-wrapper">
        <video></video>
        <div><div id="qr-code"></div></div>
        <div><div id="status-text"></div></div>
      </div>
      <div>To show/hide debug overlay press up/down or button beneath</div>
      <div id="debug-wrapper">
        <div>
          <div class="button" id="debug-button">Debug Overlay</div>
        </div>
        <div id="debug"></div>
        <div id="log"></div>
      </div>
    </div>
    <div id="info-overlay"></div>
    <script src="../lib/mozilla/object-keys-polyfill.js"></script>
    <script src="../lib/stefanpenner/es6-promise.min.js"></script>
    <script src="../lib/player.js"></script>
    <script src="../lib/manifest_parser.js"></script>
    <script src="../lib/mpd-parser.js"></script>
    <script src="../lib/wave-service.js"></script>
    <script src="../lib/davidshimjs/qrcode.js"></script>
    <script src="../lib/dpctf-testharness.js"></script>
    <script src="../lib/hbbtv.js"></script>
    <script>
      // Global variables
      var TEST_INFO = {
        id: "segment-validation",
        title: "Segment Validation",
        description:
          "Test if segments exists on server side with HTTP HEAD requests.",
        code: "segment-validation.html",
      };

      var video = document.querySelector("video");
      var qrCode = document.getElementById("qr-code");
      var statusText = document.getElementById("status-text");

      var videoContentModel = ["/content/cfhd_sets/14.985_29.97_59.94/t2/2023-09-01/stream.mpd"];
      var audioContentModel = [];

      var dpctfTest = new DpctfTest({
        testInfo: TEST_INFO,
        videoContentModel: videoContentModel,
        audioContentModel: audioContentModel,
        videoElement: video,
        qrCodeElement: qrCode,
        statusTextElement: statusText,
        infoOverlayElement: document.getElementById("info-overlay"),
        executeTest: executeTest,
        usePlayout: false,
      });

      function executeTest(player, done, parameters) {
        var minBufferDuration = parameters.minBufferDuration / 1000;
        player.startBuffering();

        let mpd;
        getMPD();

        let baseUrl = videoContentModel[0].split("/").filter(e => !e.includes(".mpd")).join("/") + "/";
        let segmentList = mpdParser.parse(mpd);
        // adjust extracted uris that they can be requested
        let segmentUri = segmentList.playlists[0].segments.map(e => e.uri).map(e => baseUrl + e);
        let segmentDuration = segmentList.playlists[0].segments.map(e => e.duration);

        dpctfTest.asyncTest(function (test) {
          var query = location.search.replace(/\?/, "");
          var match = query.match(/redirect_time=([^&]+)/);
          var REDIRECT_TIME = match ? match[1] : null;
          if (!REDIRECT_TIME) REDIRECT_TIME = 5;

          let testPass = true;
          for (let i = 0; i < segmentUri.length; i++) {
            fetch(segmentUri[i], {method: "HEAD"})
            .then(res => {
                if (String(res.status).startsWith("4") || String(res.status).startsWith("5")) {
                  console.log("Segment not exist!");
                  testPass = false;
                } else {
                  console.log("Segment exist");
                }
            });
          }

          assert_true(testPass);
          test.done();

        }, "Segment exist on Server");


        dpctfTest.asyncTest(function (test) {
          var query = location.search.replace(/\?/, "");
          var match = query.match(/redirect_time=([^&]+)/);
          var REDIRECT_TIME = match ? match[1] : null;
          if (!REDIRECT_TIME) REDIRECT_TIME = 5;

          

          assert_true(true);
          test.done();

        }, "Segment duration correct");
        
        player.playOnBufferLoaded(minBufferDuration).then(done);
      }

      // save MPD file as utf-8 string
      async function getMPD() {
        mpd = await fetch(videoContentModel[0])
        .then(res => res.text());
        mpd = JSON.parse( JSON.stringify( mpd ) );
      }
    </script>
  </body>
</html>
